# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de homologaciones y dispositivos",
        "title": "LBTech API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@lbtech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/test-cases": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Obtiene la lista de todos los casos de prueba",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Listar casos de prueba",
                "responses": {
                    "200": {
                        "description": "Lista de casos de prueba",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.TestCase"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Crea un nuevo caso de prueba en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Crear nuevo caso de prueba",
                "parameters": [
                    {
                        "description": "Datos del caso de prueba",
                        "name": "testCase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TestCase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Caso de prueba creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Caso de prueba duplicado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/test-cases/fileUpload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Procesa y carga casos de prueba desde un archivo",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Cargar archivo de casos de prueba",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Archivo de casos de prueba",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Casos cargados exitosamente",
                        "schema": {
                            "$ref": "#/definitions/responses.TestCaseFileUpload"
                        }
                    },
                    "400": {
                        "description": "Formato de archivo inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/test-cases/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Obtiene un caso de prueba específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Obtener caso de prueba por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del caso de prueba",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caso de prueba encontrado",
                        "schema": {
                            "$ref": "#/definitions/responses.TestCase"
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Caso de prueba no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Actualiza un caso de prueba existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Actualizar caso de prueba",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del caso de prueba",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados del caso",
                        "name": "testCase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TestCase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caso actualizado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Caso no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto con datos existentes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Elimina un caso de prueba del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Eliminar caso de prueba",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del caso de prueba",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caso eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Caso no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/test-cases/{id}/upgrade": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Actualiza la versión de un caso de prueba",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Cases"
                ],
                "summary": "Actualizar versión de caso de prueba",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del caso de prueba",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la nueva versión",
                        "name": "testCase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.TestCase"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Versión actualizada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Caso no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto con datos existentes",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.TestCase": {
            "type": "object",
            "required": [
                "category_id",
                "code",
                "description",
                "expected",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "expected": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "responses.TestCase": {
            "type": "object",
            "required": [
                "code",
                "description",
                "expected",
                "name"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expected": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "test_category": {
                    "type": "string"
                }
            }
        },
        "responses.TestCaseFileUpload": {
            "type": "object",
            "properties": {
                "invalid_rows": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}